added code to stop kill mode using a switch

+++ robot_drive_offense/robot_drive_offense.ino
@@ -16,6 +16,8 @@ RoboClaw rightMotors(&Serial3, 100);
 //Shooty motors
 #define LEFT_SHOOTY_PIN 6
 #define RIGHT_SHOOTY_PIN 7
+#define STOP_SUPER_FAST_SHOOTY_PIN 52
+#define STOP_FAST_SHOOTY_PIN 53
 //SoftwareSerial(rxPin, txPin)
 //BTS7960(rPwmPin, lPwmPin, enPin, invert)
 //BTS7960 leftShooter(4, 3, 7);
@@ -64,6 +66,8 @@ void setup() {
   configPwm();
   pinMode(RIGHT_SHOOTY_PIN, OUTPUT);
   pinMode(LEFT_SHOOTY_PIN, OUTPUT);
+  pinMode(STOP_SUPER_FAST_SHOOTY_PIN, INPUT_PULLUP);
+  pinMode(STOP_FAST_SHOOTY_PIN, INPUT_PULLUP);
 
   // Servo motors
   servoDriverSerial.begin(9600);
@@ -256,7 +260,7 @@ void controlScoop() {
       }
     } else {
       Serial.println("Setting scoop state: STOPPED");
-      stepper.disable();
+      stepper.disable(); 
     }
   }
 
@@ -275,6 +279,17 @@ void controlShooter() {
   static bool pressStarted = false;
 
   // Handle button presses
+  int shooterRestrictMode = 0;
+  int pinA = digitalRead(STOP_SUPER_FAST_SHOOTY_PIN);
+  int pinB = digitalRead(STOP_FAST_SHOOTY_PIN);
+  if(!pinA) {
+    shooterRestrictMode = 1;
+  } else if (!pinB) {
+    shooterRestrictMode = 2;
+  } else {
+    shooterRestrictMode = 0;
+  }
+  
   if (controller.buttonClick(DOWN)) {
     Serial.println("Setting shooter speed: OFF");
     curMode = 0;
@@ -284,7 +299,7 @@ void controlShooter() {
   } else if (controller.buttonClick(UP)) {
     Serial.println("Setting shooter speed: MEDIUM");
       curMode = 2;
-  } else if (controller.buttonClick(RIGHT)) {
+  } else if (controller.buttonClick(RIGHT) && shooterRestrictMode < 2) {
     if (curMode == 3 && !pressStarted) {// Start of a press to go into kill mode
       pressStartTime = millis();
       pressStarted = true;
@@ -294,7 +309,7 @@ void controlShooter() {
       pressStarted = false;
     }
     setShooterSpeed(speeds[3]); //high
-  } else if (!controller.button(RIGHT) && pressStarted && millis() - pressStartTime > killModeTime && curMode == 3) { // End of a press to go into kill mode
+  } else if (!controller.button(RIGHT) && pressStarted && millis() - pressStartTime > killModeTime && curMode == 3 && shooterRestrictMode == 0) { // End of a press to go into kill mode
     Serial.println("Setting shooter speed: DEATH");
     curMode = 4;
     pressStarted = false;
